fix logo onClick
